+++ HISTORY.rst
@@ -17,9 +17,10 @@ Unreleased Changes
 * Fixed an issue where ``create_raster_from_vector_extent`` would use an
   approximation of the layer extent to infer the input vector extent. Now
   uses the individual geometry to infer vector extents.
+* Fixed a bug in ``routing.distance_to_channel_mfd`` where pixels that drained
+  off the edge of the raster had negative values rather than nodata values.
 * Expanded the error message raised by ``transform_bounding_box`` when the
   bounding box cannot be transformed to provide more helpful details.
-
 
 2.2.0 (2020-05-14)
 ------------------